UWebcam: Fix FrameAdjust
The brightness scales the Y component, the hue rotates the (U,V) vector,
and the saturation scales the (U,V) vector.

Note that we could extend this to also do the conversion from BGR to RGB
for free. Converting to gray by using the 1x3 matrix [1, 0, 0] instead
of YUV2BGR should even save time in addition to making the cvCvtColor
calls in FrameEffect obsolete.

I chose to use cvCreateMat and cvSetReal2D instead of allocating the
matrices on the stack and directly manipulating their elements because
this way we don't have to worry about our CvMat declaration being correct.
The performance impact is negligible.