Enhancement of the pitch detection algorithm.
In the code it was suggested that an autocorrelation of the microphone signal was used to determine the pitch. However, an average magnitude difference (AMDF) was implemented, which is similar, but not the same.

AMDF is suitable for real-time pitch detection because it does not involve multiplications as compared to the calculation of the autocorrelation function.

An improvement of the previously implemented AMDF is the Circular AMDF, which uses the full sample window (wrapping the around to the samples at the beginning, hence circular) for pitch detection.

CAMDF suffers from double pitch errors, but since USDX ignores the octave, this poses no problem.

See http://www.utdallas.edu/~hxb076000/citing_papers/Muhammad%20Extended%20Average%20Magnitude%20Difference.pdf for more details on (C)AMDF.