opencv_highgui: Prefer symbols from opencv_videoio
Nowadays the video functions we are looking for can be found in the
video module. Prefer that library on *nix.

On Windows we ship a specific version of OpenCV and don't want to put
a videoio DLL from a later OpenCV version into the mix.

I have no idea why we compare library handles to 32.