Add wrapper for current OpenCV C++ API
The wrapper tries to look like the old C API, but IplImage.depth returns
different values. Our IplImage internally uses UMat, so it might
automatically benefit from OpenCL. Only the minimum is implemented to be
used by USDX. The OpenCV libraries are linked to the executable instead
of being loaded at runtime. Like our ProjectM wrapper, this wrapper
relies on --copy-dt-needed-entries to link to libstdc++.

CMake is not needed to find OpenCV, pkg-config works as well, but OpenCV
has declared pkg-config support as deprecated and does not build the .pc
file by default. The .pc file will also not list those modules that are
not part of the world module if the world module has been built.