UVideo: Fix looping of videos starting at time 0
Looping videos starting at 0s were broken since
f186fc5bffabce4406d1897c8ef667117ecd11d4

When the first frame of the loop has no PTS or a PTS of zero, the value
of fNetFrameTime will be used instead. But that value still was beyond
the end of the loop.

Fixes #520