UVideo: allow codec selection
The codecs have to be listed in the PreferredCodecs key of the Graphics
section of config.ini. They must be separated with comma and/or space.

Example:

[Graphics]
PreferredCodecs=h264_mmal mpeg4_mmal mpeg2_mmal vc1_mmal

If the list contains an unknown codec name, all possible codecs supported
by libavcodec that would not be chosen by default are printed to Error.log
when the first video is opened. Since for most formats there is only one
codec, this list is usually very short. A codec being mentioned in
Error.log does not mean that it actually works. F.ex. the libavcodec
included in Raspbian provides a vp9_v4l2m2m codec although there is no
V4L2 memory-to-memory device that can decode VP9.

Note that mpeg2_mmal and vc1_mmal require a license key to be bought and
VC-1 hardware decoding is not available on Raspberry Pi 4. It is safe to
list these codecs anyway because the code will fall back to the default
codec if the preferred codec fails to open the video stream.

Hardware-accelerated codecs that need av_hwframe_transfer_data to retrieve
the uncompressed frames are not yet supported and will be omitted from the
codec list printed to Error.log.